Write tests that cover routes and controllers
The tests that were written are integration tests. They cover the Post
controller and routes except for one line in the controller, which I do
not know how to cover. To do so, I need to make Post.find() throw
an exception in post-model.js, which is not possible since the method
does not accept anything.

The CI pipeline must be changed because we need a new environmental
variable called TEST_TYPE, which is either UNIT or INTEGRATION. Feel free
to change the name of this variable. For example, you could
call it IS_UNIT = TRUE or FALSE.