Add functionality to Subscribe button
The code will be changed once we can view the profiles of other users.
For now, the Subscribe button does not work, although it will if the
profile is not the user's. The button should not even appear to begin
with. It's only there so that I can test it out.